move solana estimateReceiveExecution tests to integration.test
