+++ ccip-sdk/src/solana/fork.test.ts
@@ -6,15 +6,11 @@ import { after, before, describe, it } from 'node:test'
 import { Wallet as AnchorWallet } from '@coral-xyz/anchor'
 import { Connection, Keypair, LAMPORTS_PER_SOL } from '@solana/web3.js'
 
+import '../index.ts' // register chains
 import { CCIPAPIClient } from '../api/index.ts'
-import { EVMChain, discoverOffRamp } from '../index.ts'
 import { ExecutionState, MessageStatus } from '../types.ts'
 import { networkInfo } from '../utils.ts'
-import {
-  ETHEREUM_TO_SOLANA,
-  FUJI_TO_SOLANA,
-  SOLANA_ESTIMATE_RECEIVER_MESSAGE,
-} from './fork.test.data.ts'
+import { ETHEREUM_TO_SOLANA } from './fork.test.data.ts'
 import { SolanaChain } from './index.ts'
 
 // ── Constants ──
@@ -22,9 +18,7 @@ import { SolanaChain } from './index.ts'
 const VERBOSE = !!process.env.VERBOSE
 
 const SOLANA_ROUTER = 'Ccip842gzYHhvdDkSyi2YVCoAWPbYJoApMFzSxQroE9C'
-const SOLANA_OFFRAMP = 'offqSMQWgQud6WJz694LRzkeN5kMYpCHTpXQr3Rkcjm'
 const ETH_MAINNET_SELECTOR = networkInfo('ethereum-mainnet').chainSelector
-const FUJI_RPC = process.env.FUJI_RPC ?? 'https://api.avax-test.network/ext/bc/C/rpc'
 
 // ── Surfpool helpers ──
 interface SurfpoolInstance {
@@ -276,92 +270,5 @@ describe('Solana Fork Tests', { skip, timeout: 180_000 }, () => {
 
       solanaWithApi.destroy?.()
     })
-  })
-})
-
-describe('Solana Devnet Estimate Fork Tests', { skip, timeout: 120_000 }, () => {
-  const ESTIMATE_MSG = FUJI_TO_SOLANA[0]!
-
-  let devnetSurfpoolInstance: SurfpoolInstance | undefined
-  let solanaDevnet: SolanaChain | undefined
-
-  before(async () => {
-    devnetSurfpoolInstance = createSurfpoolInstance({
-      network: 'devnet',
-      port: 8649,
-    })
-    await devnetSurfpoolInstance.start()
-
-    const devnetConnection = new Connection(
-      `http://${devnetSurfpoolInstance.host}:${devnetSurfpoolInstance.port}`,
-      {
-        commitment: 'confirmed',
-        wsEndpoint: `ws://${devnetSurfpoolInstance.host}:${devnetSurfpoolInstance.port + 1}`,
-      },
-    )
-    solanaDevnet = new SolanaChain(devnetConnection, networkInfo('solana-devnet'), {
-      apiClient: null,
-      logger: testLogger,
-    })
-  })
-
-  after(async () => {
-    solanaDevnet?.destroy?.()
-    await devnetSurfpoolInstance?.stop()
-  })
-
-  it('should estimate receiver execution for a failed Fuji -> Solana devnet message', async () => {
-    assert.ok(solanaDevnet, 'Solana devnet chain should be initialized')
-
-    await using disposer = new AsyncDisposableStack()
-    const source = disposer.adopt(
-      await EVMChain.fromUrl(FUJI_RPC, { apiClient: null, logger: testLogger }),
-      (source) => source.destroy?.(),
-    )
-
-    const tx = await source.getTransaction(ESTIMATE_MSG.txHash)
-    const requests = await source.getMessagesInTx(tx)
-    assert.equal(requests.length, 1, 'tx hash should contain one CCIP message')
-
-    const request = requests[0]!
-    assert.equal(request.message.messageId, ESTIMATE_MSG.messageId)
-
-    const offRamp = await discoverOffRamp(source, solanaDevnet, request.lane.onRamp, source)
-    const estimated = await solanaDevnet.estimateReceiveExecution({
-      offRamp,
-      message: {
-        sourceChainSelector: request.lane.sourceChainSelector,
-        messageId: request.message.messageId,
-        receiver: request.message.receiver,
-        sender: request.message.sender,
-        data: request.message.data,
-        tokenReceiver:
-          'tokenReceiver' in request.message ? request.message.tokenReceiver : undefined,
-        accounts: 'accounts' in request.message ? request.message.accounts : undefined,
-        accountIsWritableBitmap:
-          'accountIsWritableBitmap' in request.message
-            ? request.message.accountIsWritableBitmap
-            : undefined,
-      },
-    })
-
-    assert.ok(
-      30_000 < estimated && estimated < 33_000,
-      `estimated compute units should be around 31k for this message, got ${estimated}`,
-    )
-  })
-
-  it('should estimate receiver execution for a real message with token transfer', async () => {
-    assert.ok(solanaDevnet, 'Solana devnet chain should be initialized')
-
-    const estimated = await solanaDevnet.estimateReceiveExecution({
-      offRamp: SOLANA_OFFRAMP,
-      message: SOLANA_ESTIMATE_RECEIVER_MESSAGE,
-    })
-
-    assert.ok(
-      42_000 < estimated && estimated < 45_000,
-      `estimated compute units should be around 43k for this message, got ${estimated}`,
-    )
   })
 })
